🔵 use router and put routes into "controller" directory
